添加 tinyint, samllint, int, float 数据类型 修改数据存储格式 优化查询性能

+++ src/pinusdb/port/os_file.cpp
@@ -304,7 +304,7 @@ PdbErr_t OSFile::Read(void* pBuf, size_t count, size_t offset)
 
   _Seek(offset);
   ssize_t ret = read(handle_, pBuf, count);
-  if (ret == count)
+  if (ret == (ssize_t)count)
     return PdbE_OK;
 
   LOG_ERROR("failed to read ({}) offset({}), count({}), pBuf({}), err:{}",
